new nodes with an alias config reported as new
previously the aliasTo node initializer would pre-empt the logic
that supports the isnew flag. This would prevent parent nodes from
being updated accordingly. Since the alias contract mostly has to
do with interactions with child nodes, and explicitly does not
share meta data on the alias'd node itself, it is unintuitive for
it to not update parent nodes' data attachments and creates an
inconsistency between eg. the parent's count of child nodes and
how its data attachments are supposed to behave that we could
otherwise rely on.

Since I believe this removes the need to check the hits counter on
newly created child nodes for the once flag, that check has also
been removed.