delay bundle skipping and compression handling
currently we usually block waiting for a remote response as soon
as we try to wrap the input stream in a decompression wrapper. this
is because the constructor sometimes tries to immediately read bytes
(like for a header or something). Bundle skipping is similar in that
it _would_ block given the chance. This change tries to move any
blocking network wait until after the source is pulled in turn from
the preOpen queue.

in order to take advantage of this change, jobs will probably need
to make sure meshy.stream.prefetch is set to true, preOpen is large,
and multiBundleReads is set relatively high (to try to pull as much
available data from wraps that have already paid the initial wait -
which tends to be the longest - before hitting the new wrap)