Adding bundle canaries to Hydra jobs.
These are a new type of alerts that can be added to map jobs.
The bundle canary alerts specify a query path, query ops, query
remote ops, and a bundle filter. The query must return one or
more rows. The first row is interpreted as a header row. Each
subsequent row is interpreted as a bundle. The bundle filter
is applied to each row and the alert is fired if any of the rows
yield a value of false.

Squashed commit of the following:

commit af21e08587648c4dcbb26b80882c9e9ec001aaa4
Author: Michael Spiegel <spiegel@addthis.com>
Date:   Tue Mar 25 16:14:12 2014 -0400

    Removed unused imports.

commit 4f1e9859f1974fd82623ad12d172db647e573104
Author: Michael Spiegel <spiegel@addthis.com>
Date:   Tue Mar 25 16:05:22 2014 -0400

    Refactoring of job alert error handling.

commit 38620145543512e54bbd21bc4fdf09a52e428234
Author: Michael Spiegel <spiegel@addthis.com>
Date:   Tue Mar 25 15:48:52 2014 -0400

    Improved error handling when specifying alerts.

commit 7ca80a665601a1a7878d2f5dd84ce3cc97b294fd
Author: Michael Spiegel <spiegel@addthis.com>
Date:   Tue Mar 25 13:36:02 2014 -0400

    Initial commit of canary query filters.