Delegate hocon and json configs from TaskRunner.
Allow TaskRunner to run both hocon and json configuration. Create an
API that accepts a configuration body as input, and yields the
configuration type as output. We may need to tweak the API
over time in order to meet our needs. But having this mechanism
in place allows for the job validation and related parsing features we want for
all types of jobs.