Bugfix for empty ConcurrentKeyTopper#bytesDecode().
ConcurrentKeyTopper#bytesEncode() has a short cut to encode an empty key topper
with a zero length byte array. The bytesDecode() method did not check for this
case and was throwing a exception on attempt to decode the zero length byte array.
Added unit tests for encoding empty and non empty top-K items.