MarkDB database errors should error the job.
Previously markDB database errors were silent. They were logged in
the standard output and job continued to execute normally. Now these
errors cause the job to exit abnormally. This commit touches a lot of
files but most files have minor changes. The logic for erroring a job
is buried in TaskFeeder. I had to pass around an AtomicBoolean so that
all threads involved would be reading and/or writing the same state
to determine if a job has errored. To engage the old behavior of silent
failure set 'hydra.markdb.error.ignore' to True.