Adding parameter 'double' to reservoir query.
Allows the reservoir query to store the double value for the
mean, standard deviation, and other calculated values in the
'hits' of a node by converting the double into its bit
representation and storing that in the long. Default is
to disable this behavoir so humans can read the results of
the query without the byte conversion.