Improve 'raw' argument to reservoir data attachment.
The 'raw' parameter for the reservoir data attachment query was originally
itended for debugging purposes. It dumps out all of the elements of the reservoir
as tree nodes. This commit allows only a subset of the nodes to be returned
for space efficiency. Additionally it allows for fine grained access of raw
values by respecting the 'epoch' parameter.