catch throwables in a few more places
unfortunate but since the on out of memory error jvm flag does not
trigger on these, we need to do something about them sometimes.
this handling is only marginally better but i'll take it for now.
given that it is actually off-heap and not some donkus that isn't
killing-ing on on-heap memory errors, it isn't as impossible to
run a bit of logic.