remove 'commented list of hosts' config magic
this only triggered during job creation so only external,
programmatic job creation can possibly be affected. Such logic,
if any, should just use the hosts parameter when hitting the
submit/ save end point.

The cause for removal is the lack of added value compared to the
cost of the failure mode. Given the generic nature of the
triggering magic, it isn't even really all the unlikely to be
triggered by mistake, and the error reporting when it does is
incredibly unhelpful.