custom log4j2 config parsing via typesafe-config
can't quite use all of codec's supreme glory due to some
incompatabilities in their respective plugin systems, but this
hack goes quite a long way to wards making their "json config
syntax" more bearable. I had forgotten how obnoxious extra/
missing commas, and constant quoting, could be.

in addition to all the standard hocon advantages, I also put in
a little gizmo that is designed to make modifying the logging
config from job configs easier. if the job config is the
specified logging config file, and it has any 'global' section
defined, then it is merged with the global defaults and the
'global.logging' is plucked out to form the resolved log config.
A reference.conf is provided that defines the default 'global.
logging' as just the inclusion of log4j2.conf.