work around stupid hardcoded log4j2 key indexes
if 'properties' isn't the first key in a json object, then they
just stop working altogether. presumably a bug or at least an
unintended 'feature'. this patch adds shallow support for ignoring
top level keys that are prefixed with _s. this makes it convenient
enough to just use hocon substitution for most purposes.