new/fixed task auto-revert-retry logic
There is a new (boolean) setting for allowing automatic retries called
"autoRetry", or "Enable Automatic Retries" as seen via the spawnv2
settings page (as a check box). The contract is that if enabled, we
are allowed to make best-effort attempts to revert retry task runs in
the face of what appear to be transient errors. The exact conditions
and number of times that retries may be performed is left intentionally
unspecified, but for now it retries a maximum of one time, and only in
response to a JVM error (or other SIGABRT / 134 exit codes).

There is a new meter metric which tracks how often this occurs, but
it does not do so on a per-job basis.

This also removes the old 'retries' config setting which appears
to have had (in addition to being unused) two bugs:

- skipped the most recent backup
- did not actually retry the task, just reverted. therefore
  it reverted at most once, and could cause a job with a non-transient
  error to forever silently revert without making progress.

Additionally, this commit also removes some deprecated options from
the spawn v2 job settings tab: query trace level, query consecutive
failure threshold, read only replicas