swallow less errors in task components ref T37568
not a perfect long-term solution. That probably involves more
usage of completable futures, blocking close calls, and/or
letting sources elect to push bundles rather than use map-feeder
polling.

One clear symptom of this is that ASFDS worker threads have no
reliable way to report errors that occur while one of them is
calling shutdown, and so have no option other than to halt the
jvm. In fact it also looks like there can be a race condition
causing the jvm to finish exiting prior to shutdown completing
when a SIGTERM is handled after ASFDS has just started to exit
on its own.

However, this is a vast improvement for the short term (barring
any serious mistakes).