bump codec, jar-pom, jackson for delegate fixes
this works best on things like ValueFilterNot. it works reasonably
well for the unary bundle filters, but the delegate object is not
allowed to be optional. So `false {}` and `false {some-filter}`
can no longer be supported by the same class. The upside is that
they should now all work correctly with non-single-key delegates,
defaults, and so on.