map.fields entries more similar to b.f.Field
the `to` field can be left out when it matches the `from` field,
and the `filter` specifier can be left out similar to when using
the default bundle filter and including "extra-filters".

also note that the previous commit erroneously contained some of
the changes noted here (and in incomplete form).