ensure Streams obtained via Files.list are closed
there is some noise in reference/ documentation materials about
terminal operations "closing" streams, and it is certainly possible
to do programmatically with the Stream api. However, this is a
different sort of "close" and is only related in that it causes the
same exceptions. They cannot be disambiguated.

given that they could easily wrap all the terminal operation methods
in try/finally/close style blocks, it is somewhat surprising that
they chose not to. It seems that in the mailing list they decided
that "AutoCloseable" is a flawed interface, not their problem, and
that it was okay to make it substantially less useful by routinely
using it as a no-op -- and _then_ encouraging people not to err on
the side of caution. They have lost their collective minds.

in any event, this change corrects a previous misunderstanding on
my part.