Support for matching entire level in path query.
In the 'query' path operation the current functionality is to
return a single matching node. This patch extends that functionality
with the ability to return multiple matching nodes. An entire
level of the tree can be returned by the type "glob" path element.
"glob" can only be used in the context of a query path operation.