many fixes and refactors for mesh source
when a task was stopped early (either via run time or manual use
of the stop button):
- the JVM could exit before the source had finished exiting
  correctly. This could leave the marks db in an inconsistent
  state and other "bad things".
- bundles could get dropped by being added after the "term
  bundle"
- if the stop occured early in source setup, it could get swallowed
- when a source was extremely slow to produce bundles, it could be
  just as slow to respond to the sig term

in general, the close method is now a blocking call supported by
the new ability of worker threads to hold onto a single bundle
without committing to it being read / added to the queue. This
means that all worker threads will exit before the close method
returns.

error reporting is now done in the worker threads by simply throwing
exceptions. They are handled from the task futures. The task futures
are also used to track when the worker threads (tasks really) have
finished rather than a custom count down latch.

the "term bundle", several state fields, and methods were simplified
away. I spent most of the LOC gained this way on extra spacing and
comments, but there should still be a decent net negative. The
usual basics like making fields final where trivial, and so on are
also included.