Filter initialization is now explicit. Ref T51853.
Bundle filters and value filters have a lazy initialization
mechanism that is poorly implemented and a bad design.
A developer writing a new filter must remember to
explicitly call the initialization method and it is
also possible to invoke other filters and bypass
the initialization.

In this implementation all filters have an open()
method. The application is responsible for ensuring
that open() is called before using the filter.

In this implementation the majority of initialization
is be performed in the constructor. Any initialization
that is too expensive to perform when the object
is deserialized (such as accessing the network or
the filesystem) should be placed in the open() method.
The use of SuperCodable for filters is discouraged.