Allow PathQuery#childMatch to return map of values.
Replace the PathQuery#childMatch boolean flag with an enum field.
The existing behavior will return a list of values if childMatch is true.
In the new behavior childMatch can return either a list or a map.
Providing this feature speeds up my job by x5 for its particular
use-case of childMatch by allowing for the deletion of a specific
key without iterating through all the elements.