Change constructor argument name for documentation builder.
The documentation builder generator currently looks at the name of the
arguments not the @JsonProperty value to match against field names.
This is a bug and should be fixed (I tried and couldn't retrieve the
value from within the @JsonProperty). In the interim changing the
argument name so that it will appear in the generated documentation.