Move PathPrune expensiveShutdownTest up call tree.
The expensiveShutdownTest() is only used because the order of the
shutdown hooks may allow state.processorClosing() to return false
for a while when the JVM is shutting down. When we call
the expensiveShutdownTest() does not change the correctness of
the algorithm so it can be moved outside all the iterators.