Job alert for discrepancies in # of log files per task.
Refactored the job alert from a single class with hand made
enumerated values into a class hierarchy, with one class per
alert type. Create LogCountJobAlert that looks for any high
or low deviations in the number of log files across the tasks
of a job. Implemented wildcard notation in job alerts. If a single
'*' is used to specify the job identifiers then the alert is
applied to all jobs in the cluster. We scan through at most N jobs
(N = 50 default) for each invocation of the wildcard. Any active
alerts are always tested on each iteration as consistent with the
current behavior.