replace "narrow" annotation with primitive field
this costs a few more in-memory bytes per node, but it is still p much an
insignificant cost compared to existing fields (especially for concurrent
nodes). This also removes the memory indirection of a wrapping object and
restores a lot of symmetry with the serialized form. We could obviate the
need for this by using legacy shell object kinda like we do for old marks
entries, but for in-memory-only costs, there are are lot of lower hanging
fruit anyway.

while I was at it, I went ahead and removed the 'volatile' flag since it
does not appear to be required.