fix perf. regression + simplify VFRequire/Exclude
this uses some new features in codec/bundle that allow for a
"TypedField<T>" field / parameter declaration. This type preserves
performance for constant types that have non-basic attributes.

In particular, it allows us to preserve the abstraction for the end
user without giving up control over the representation of the collection
of some strings where being a set instead of a list is important for
performance. Much cleaner inter-op with normal java code as well as a
bonus.