Wildcard job alerts must always test active alerts.
Wildcard job alerts (job alerts that test all the jobs in the cluster)
only process a subset of all jobs on each iteration of scanning
alerts. This is to prevent the job alert thread from consuming
an arbitrary amount of CPU time. Since they process a different subset
on each iteration, we must make sure that in addition they reprocess
any active alerts. Otherwise we will make the active alerts as not fired.
This is not an issue for regular job alerts which always process all
jobs that apply to them on each iteration. Ref T52172.