Spawn: add mapping from job ID to set of alert IDs.
This mapping is used to clean up alerts when a job is deleted.
If alert to be updated is watching only the deleted job then
delete the alert. Otherwise save a new alert that removes
the deleted job ID. Job aliases are not de-referenced into
the corresponding job ID. Alias deletion does not clean up
alerts. This can be implemented in a future commit if necessary.
Unit tests added for alert cleanup on job deletion. Ref T53179.