fix require/exclude exact matching on non-strings
before TypeField was introduced, this had been switched to do a
contains operation using the base value-object because we could
easily get a Collection<ValueObject> but not a Collection<String>.

after we switched back to using Collection<String> (specifically so
we could have a Set<String>), we forgot to change back the argument
type. This passed all tests because ValueStrings will unfortunately
play it fast and loose with the equals method contract and represent
themselves as equal to strings, and apparently the contains method
only calls equality methods in that one direction.

additionally, this error type was not being reported correctly by
intellij due to some issues related to custom severities. That has
been resolved and so this will hopefully not go unnoticed again.