Implement pipeline job type.
A pipeline job will run the first job phase to completion, then the second
phase, then the third phase, etc. When a pipeline job stops and is restarted
it begins processing from the first phase. Added error checking to verify
that each job phase does not write to an output directory of another job phase.
The support for conflicting output directories can be used to backport
this feature to ordinary hydra jobs with conflicting mark dirs or
conflicting outputs. Ref T53265.