Cleanup tree trash before 'pre' and after 'post'.
The cleanup operation will terminate if the JVM is shutting down.
Previous implementation was cleaning up only before the 'pre' operation.
This ensures that cleanup has first priority on the available time for a job.
In this commit cleanup also happens after the post operation. If the 'pre'
or the 'post' phase generated any garbage then we now attempt to cleanup
the trash at the end of a job if there is avaiable time. Ref T52172.