Force task close exceptions to error the JVM.
This change is more controversial is it will also effect all
types of jobs. However it would seem to fall into the category
of a bug fix as the initialization of a task causes a non-zero
exit status. It follows that the closing of a task should do the
same.