Improve error handling of WriteableDiskCheck.
Addresses the issues that were brought up in https://github.com/addthis/hydra/pull/131. Add a wrapper class that executes child processes.
Does not attempt to consume from standard output or standard
error. Kills the child process if the child is blocking
on the output streams. This is preferred for our use cases
where the alternative is to create an extra thread to consume
from these streams.