fix default cacheWeight option
I had gone through a couple variations on what type of number to use,
and it seems I got mixed up in the end. (integers are typically
easier to wrangle than floats, but I wanted to be able to opt for a
smaller cache size, and making "100" the baseline seemed more likely
to be confusing to end users).