retry netty 4.0.26 with work-around
work-around was implemented in meshy. the bug we are aware of does
not seem to be much of a risk for impacting the query http server
cause we don't do much init until we get an http request, and we
don't do any client connections ourselves. Therefore, we have a
decent chance of being okay.