remove TreeMapState.leases + fix nested elements
the responsibility to release nodes has been moved from a per-bundle
garbage collector to the caller of methods that return them. This
greatly decreases the number of concurrently leased tree nodes in
many cases, although it would be nice if we knew in advance that the
collective leaf nodes were not interesting to the root caller (this
can be effected with a path element whose filter always fails placed
"after" the leaf nodes, but such an arrangement is inconvenient).

this also fixes an apparent bug in the processing of sub-groups of
path elements -- such as those in PathBranch, PathCall, and PathChild.
If the last path element in such a sub-group generates multiple nodes
-- such as a PathValue operating on a value array -- then subsequent
child nodes occuring below the sub-group would only be generated for
the last parent node to be processed. eg. `{field_a: [a, b], field_b:
 c}` could result in the top level nodes `a` and `b`, but only `b`
 might have a child node `c`.