Add LeasedTreeNodeList abstraction.
The purpose is to simplify the node leasing contract by moving all the
logic into a central location. The collection will ensure that either
its contents are read exactly once or it is released exactly once.
The guarantees are weak in that if the library developer forgets to
invoke release we don't do anything other then issue some logging messages
and fail on an assertion and these actions happen in a finalizer (bleh).