Simpler query input source. Uses mqmaster.
Submits queries to the mqmaster. The mqmaster is used
to allocate cpu resources fairly across these queries
and regular queries to the query system. Each query will only
submit to a subset of tasks. The queries submit only
rops there are no ops. A markfile is used to signal that the
input source finished reading the query and that the query will
not execute on rekick. In this implementation the job will error
if partial query results are read. This is a temporary solution
until a comprehensive solution for reading partial queries is constructed.

Ref T53270. Ref T53475.