Use streaming match for /+a,b,c,.. type queries.
A very synthetic benchmark shows a x1.7 performance improvement
with this patch. The benchmark uses the following job
http://pastebin.com/SKbHBEnq. Its not advisable to have the job
run to completion it can be stopped after generating a few million nodes.
Then a query is generated with the path "/date/+0000000,0000002,0000004,.../+"
where the last node in that ... series is 2000000. This query is submitted
via curl to the query master running on local stack. It is submitted several
times to ensure that the query cache does not change performance. Completed
in 60 seconds with patch and 100 seconds without patch.