Query notation for specifying a default value.
There was a feature request to implement regular expression type
matching in the query path syntax for repeated expressions of
type X{n}, X{n,}, X{n,m}. This is a simpler form of that feature.
It implements a query path notation of the form "/+[foo]" or
"/+[foo]a,b,c". The string enclosed in the square brackets is a
default value that is returned when no matches are found. The
same notation works for the data attachment queries "/+%[foo]name=parameters"

There are several benefits to this approach. It is simpler
than the original request and was implemented quickly. It has
a more straightforward semantics in the sense that it ensures that
a column is always generated for the "+" request. It was unclear
whether there would be rows with variable-length columns in the
original X{n,m} request. It also meets the use case of the original
feature request.

Ref T52644.