Remove unused parameters to page and node iterators.
Removed the 'sampleRate' and 'inclusive' parameters. The
sampleRate parameter is unused (defualt is 1). The inclusive parameter
was unused in any external request (default is true). The only
use of the inclusive parameter is in an implementation detail of the
SkipListCache and it has been preserved for that instance.
Any external facing requests always request the lower bound to be inclusive
(ie. = true).