Adding support for bypassing the dateFormat.
This is something of a hack but it is an improvement upon the current
design in that it explicitly adds a mechanism for bypassing the date
range scanning in AbstractPersistentStreamSource. The previous
best practice was to "fake" the date range which is now causing problems
with the new date validation. Eventually this functionality
should be moved into its own class.