Spawn quiesces if rabbitmq cannot emit messages.
The rabbitmq network enters an unhappy state if messages are being
produced and not consumed. But the cluster is in a very bad state
when messages can no longer be produced. When messages are dropped
on the floor the cluster can enter an inconsistent state. This
automatically quiesces the cluster to minimize the damage. Ref T54274.

This pull request branches from https://github.com/addthis/hydra/pull/168