truncate log lines over 1000 characters (as tail)
While minion should be smart enough to handle this on its own, there is
little value in letting tasks write lines larger than this anyway. This
uses a log4j setting that truncates the "message" component to a maximum
character limit, but it does so by clipping the start of the message.

There is a patch for the next version to provide the more sane option of
clipping the end of the message instead (although no mention of the
ideal -- clip out the middle), but this is better than nothing.

fixes T54627