mv ...hydra.job.minion package to ...hydra.minion
The com.addthis.hydra.job package is probably mislabeled.
It is the umbrella package for all job management including spawn.
All of the subpackages until com.addthis.hydra.job are primarily spawn and
job scheduling related, with the exception of com.addthis.hydra.job.minion
which is responsible for the minion.