remove OpDiff from the user-reference
its docs are insufficient to describe how it is used (and reading the
code does not even help in this case). The name alone may be misleading,
so while we will leave it available for any possible legacy usages, we
don't want people attempting to use it from the user ref.