Futureproof against rollback of https://github.com/addthis/hydra/pull/174
When PR 174 is merged into master, if the hydra release needs to
be reverted after the merge, then this commit will ensure that
new fields will be ignored. This commit also allows test environments
that alernate between master and branches to continue to work.
These annotations must be dropped when PR 174 is merged to master.