Add 'noAdminUsers' field to static authentication.
Add a blacklist of users that cannot be administrators. This is
convenient when bots are members of a group that has admin priviledges
and we do not want the bots to have admin priviledges.

Reduced visibility of adminGroups() and adminUsers() methods. They
are intended only for the isAdmin() method and visibility has been
changed from package-private to protected to enforce this usage.