Setting default authorization mode to "basic".
Default authentication remains at "allowall". So you can type
anything you want for your username and password. Changing
default authorization to "basic" requires that the sudo button
be pressed for administrative actions. This configuration
is similar enough to a production environment to be useful
in local testing, and permissive enough such that any username
and password are accepted.