/job/save API allow job defaults to be optional.
Previously an arbitrary subset of job parameters were assigned default
values in the Spawn backend when a new job was created. Preserve this
behavior for legacy applications and update the Spawn UI such that
defaults are no longer assigned. This fixes the Spawn UI glitch where
some job parameters would magically be populated after a job is saved.
It also allows for the overriding of default true values with a value
of false.