variety of fixes for minion task state recovery
in particular, this resolves at least three different bugs that were
preventing minion from recovering from termination while a task was in
the 'replicating' or 'backup' state. There are probably still some edge
cases, but in general, bouncing minions with non-idle tasks should be
much more likely to be well-handled.