PathPrune: treePath and allLeaves parameters.
New 'allLeaves' parameter which ignores the date formatting behavior
and will delete all leaf nodes in the current subtree. Added unit
test for 'allLeaves' and verified that existing unit tests continue
to work (ie the existing tests have not started to delete all data).
New feature '{{DATE}}' to the treePath that will perform date
matching in the middle of a path.