Do not save task state on work item "clear"
this leaves a serialized form of the task which is in an indeterminable
state -- the intention to immediately start another work item phase
represented in the code is lost when minion exits. By not saving here,
we plan for recovery to simply reclear and then pick up where it left
off.