Add treeTrashNode to non concurrent tree.
Allows for responsive shutdown of a tree. If a node with children is
deleted it is moved into the trash. The trash is cleaned out when the
tree is closed, assuming the task is not terminating, or the trash
is always cleaned with the task is started.

This implementation reintroduces the behavior of using the call stack
to follow tree depth. The implementation is simpler when we have to
account for task shutdown during deelteSubTree(). If you feel strongly
about the possibility of stack overflow we can change the implementation.