Treat quiesce as threshold and keep kick priority
There are a couple complex cases where I have yet to bother to find the
best way to make sure the per-kick priority is either maintained or
given the most sensible value possible. Some of these are probably
correct as is, but for instance, there is a case where a task is
rebalanced while queued, and since the task object does not imply easy
access to the queue item that encloses it, the priority is not trivial
to recover there.

That difficulty will be more generally solved in a subsequent change
since it will be needed when deciding whether tasks can be preempted.

fixes T58436
fixes T58441