Performance optimization for PathPrune.
If nameFormat based pruning is occuring then we can stop looking at
sibling nodes afte the first node that is not deleted. This is because
the nodes are stored in sorted order. Unfortunately this optimization
has to be opt-in because the nameFormat may not be the lexicographic
order of the nodes (although in practice it is likely to be the
order).