Attempt to be explicit about query node filters
Previously, we relied on whether or not the node name, or any
sub-property or data attachment, should be visible in the output to
decide whether or not we cared about processing any specified and what
the resultant possible reported modifications should be handled as. This
was problematic in cases where we desired to indicate no intention of
modifying the bundle itself via adding columns, but still wanted to be
able to cut off processing or other arbitrary operations.

The calling code previously treated "0"s as implied instances of
the former case. "0" is a reasonable value where we might want the
latter case, but choose not to act on it, so that flag was simply moved.

fixes #T59934